execution_service/requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
drf-spectacular==0.27.0
python-decouple==3.8
requests==2.31.0


